# Avoid copy of flipped A matrices in GEMV