Related #1388 

Benchmark results for the new test

## Before
```
------------------------------------------------------------------------------------------------------------------- benchmark: 8 tests ------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                            Min                   Max                  Mean              StdDev                Median                 IQR            Outliers          OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_gemv_negative_strides_perf[F_layout-pos_stride0-pos_stride1]        42.4400 (1.0)        237.6250 (1.45)        74.0318 (1.36)      10.5868 (1.23)        70.1510 (1.37)       8.2260 (2.72)      887;607  13,507.7001 (0.73)       7481           1
test_gemv_negative_strides_perf[C_layout-pos_stride0-pos_stride1]        49.3120 (1.16)       164.1970 (1.0)         54.3276 (1.0)        8.6408 (1.0)         51.2660 (1.0)        3.0282 (1.0)      610;1045  18,406.8406 (1.0)        7529           1
test_gemv_negative_strides_perf[C_layout-neg_stride0-neg_stride1]       177.6830 (4.19)       663.7110 (4.04)       234.9203 (4.32)      60.4707 (7.00)       210.0830 (4.10)      47.0472 (15.54)     460;329   4,256.7635 (0.23)       3313           1
test_gemv_negative_strides_perf[C_layout-neg_stride0-pos_stride1]       188.0110 (4.43)       824.5320 (5.02)       255.6154 (4.71)      84.1727 (9.74)       241.2860 (4.71)      73.3770 (24.23)     219;158   3,912.1266 (0.21)       2386           1
test_gemv_negative_strides_perf[C_layout-pos_stride0-neg_stride1]       246.1110 (5.80)       729.8850 (4.45)       329.7730 (6.07)      75.7956 (8.77)       299.0840 (5.83)      71.1830 (23.51)     458;245   3,032.3893 (0.16)       2828           1
test_gemv_negative_strides_perf[F_layout-pos_stride0-neg_stride1]     1,460.7620 (34.42)    3,241.3030 (19.74)    1,688.4473 (31.08)    114.0200 (13.20)    1,686.8400 (32.90)     90.3190 (29.83)      107;50     592.2601 (0.03)        600           1
test_gemv_negative_strides_perf[F_layout-neg_stride0-pos_stride1]     1,463.9280 (34.49)    1,823.5010 (11.11)    1,633.8600 (30.07)     97.6424 (11.30)    1,685.4870 (32.88)    180.8180 (59.71)       218;0     612.0475 (0.03)        574           1
test_gemv_negative_strides_perf[F_layout-neg_stride0-neg_stride1]     1,468.8780 (34.61)    1,826.1960 (11.12)    1,699.7320 (31.29)     66.4291 (7.69)     1,711.2460 (33.38)     51.0550 (16.86)      111;63     588.3281 (0.03)        582           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

```

## After
```
------------------------------------------------------------------------------------------------------------- benchmark: 8 tests ------------------------------------------------------------------------------------------------------------
Name (time in us)                                                         Min                 Max               Mean             StdDev             Median                IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_gemv_negative_strides_perf[C_layout-pos_stride0-pos_stride1]     48.5710 (1.0)      183.2230 (1.07)     53.4605 (1.0)      10.4242 (1.23)     50.2840 (1.0)       2.7525 (1.27)     507;1059       18.7054 (1.0)        7419           1
test_gemv_negative_strides_perf[F_layout-neg_stride0-neg_stride1]     57.3480 (1.18)     191.2980 (1.12)     76.0201 (1.42)     13.1701 (1.56)     78.7670 (1.57)     18.5050 (8.51)     2476;161       13.1544 (0.70)       7486           1
test_gemv_negative_strides_perf[F_layout-pos_stride0-neg_stride1]     58.5190 (1.20)     850.2520 (4.96)     89.4035 (1.67)     22.1240 (2.61)     83.6570 (1.66)      9.5490 (4.39)      578;874       11.1852 (0.60)       8390           1
test_gemv_negative_strides_perf[C_layout-neg_stride0-neg_stride1]     59.7120 (1.23)     392.5150 (2.29)     69.6952 (1.30)     16.4890 (1.95)     62.2770 (1.24)     10.9200 (5.02)      756;677       14.3482 (0.77)       6077           1
test_gemv_negative_strides_perf[C_layout-pos_stride0-neg_stride1]     60.0820 (1.24)     182.9930 (1.07)     64.4955 (1.21)      8.4641 (1.0)      61.6260 (1.23)      2.5850 (1.19)     520;1053       15.5050 (0.83)       7272           1
test_gemv_negative_strides_perf[C_layout-neg_stride0-pos_stride1]     61.4150 (1.26)     171.5520 (1.0)      67.0458 (1.25)      9.7855 (1.16)     63.3090 (1.26)      4.3470 (2.00)      462;630       14.9152 (0.80)       5358           1
test_gemv_negative_strides_perf[F_layout-pos_stride0-pos_stride1]     67.1760 (1.38)     236.6830 (1.38)     78.6778 (1.47)     21.5318 (2.54)     72.0900 (1.43)      8.8660 (4.08)      541;734       12.7101 (0.68)       7472           1
test_gemv_negative_strides_perf[F_layout-neg_stride0-pos_stride1]     78.0960 (1.61)     203.8410 (1.19)     84.2026 (1.58)     13.0079 (1.54)     79.5090 (1.58)      2.1740 (1.0)      550;1251       11.8761 (0.63)       6668           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

There are still other things we probably want to do, regarding the copy / allocation layout of y when not inplace, but that I won't touch now.

We can do a similar trick for the GER implementation, and perhaps other blas Ops

<!-- readthedocs-preview pytensor start -->
----
📚 Documentation preview 📚: https://pytensor--1389.org.readthedocs.build/en/1389/

<!-- readthedocs-preview pytensor end -->